EY-4942 Slå sammen PdlTjensterKlienter
