+++ apps/etterlatte-behandling/src/main/kotlin/config/ApplicationContext.kt
@@ -101,10 +101,12 @@ import no.nav.etterlatte.funksjonsbrytere.FeatureToggleProperties
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggleService
 import no.nav.etterlatte.grunnlag.GrunnlagHenter
 import no.nav.etterlatte.grunnlag.GrunnlagService
+import no.nav.etterlatte.grunnlag.IOpplysningDao
 import no.nav.etterlatte.grunnlag.OpplysningDao
 import no.nav.etterlatte.grunnlag.OpplysningDaoProxy
 import no.nav.etterlatte.grunnlag.aldersovergang.AldersovergangDao
 import no.nav.etterlatte.grunnlag.aldersovergang.AldersovergangDaoProxy
+import no.nav.etterlatte.grunnlag.aldersovergang.IAldersovergangDao
 import no.nav.etterlatte.grunnlagsendring.GrunnlagsendringsHendelseFilter
 import no.nav.etterlatte.grunnlagsendring.GrunnlagsendringshendelseDao
 import no.nav.etterlatte.grunnlagsendring.GrunnlagsendringshendelseService
@@ -294,6 +296,8 @@ internal class ApplicationContext(
             env.requireEnvValue(SKJERMING_URL),
         ),
     val brukerService: BrukerService = BrukerServiceImpl(pdlTjenesterKlient, norg2Klient),
+    val opplysningDaoProxy: IOpplysningDao = OpplysningDaoProxy(config, httpClient()),
+    val aldersovergangDaoProxy: IAldersovergangDao = AldersovergangDaoProxy(config, httpClient()),
 ) {
     val httpPort = env.getOrDefault(HTTP_PORT, "8080").toInt()
     val saksbehandlerGroupIdsByKey = AzureGroup.entries.associateWith { env.requireEnvValue(it.envKey) }
@@ -356,24 +360,21 @@ internal class ApplicationContext(
     val skalBrukeDaoProxy = true
     val opplysningDao =
         if (skalBrukeDaoProxy) {
-            OpplysningDaoProxy(config, httpClient())
+            opplysningDaoProxy
         } else {
             OpplysningDao(dataSource)
         }
 
-    val grunnlagPdlTjenester =
-        no.nav.etterlatte.grunnlag.klienter
-            .PdlTjenesterKlientImpl(pdlHttpClient(config), config)
     val nyGrunnlagService: GrunnlagService =
         no.nav.etterlatte.grunnlag.GrunnlagServiceImpl(
-            grunnlagPdlTjenester,
+            pdlTjenesterKlient,
             opplysningDao,
-            GrunnlagHenter(grunnlagPdlTjenester),
+            GrunnlagHenter(pdlTjenesterKlient),
         )
 
     val aldersovergangDao =
         if (skalBrukeDaoProxy) {
-            AldersovergangDaoProxy(config, httpClient())
+            aldersovergangDaoProxy
         } else {
             AldersovergangDao(dataSource)
         }
